% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstream.R
\name{cf_upstream_ipv4}
\alias{cf_upstream_ipv4}
\alias{cf_upstream_ipv6}
\title{Cloudflare upstream source ranges used to query upstream DNS providers}
\usage{
cf_upstream_ipv4()

cf_upstream_ipv6()
}
\description{
Cloudflare upstream source ranges used to query upstream DNS providers
}
\note{
makes a network connection to Cloudflare to retrieve the list.
    Consider 'memoise'-ing it if you use it in "production".
}
\examples{
cf_upstream_ipv4()
cf_upstream_ipv6()
}
