% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnsflare-package.R
\docType{package}
\name{dnsflare}
\alias{dnsflare}
\alias{dnsflare-package}
\title{Query 'Cloudflare' Domain Name System ('DNS') Servers over 'HTTPS'}
\description{
Domain Name System ('DNS') queries are generally perfomed by operating
system calls when browsing the internet or making other network connectios. DNS
over 'HTTPS' services offer more local confidentiality and some 'DNS' providers
like Cloudflare offer (\url{https://developers.cloudflare.com/1.1.1.1/dns-over-https/json-format/})
an 'API'-like service along with other privacy-enabled 'DNS' offerings. Tools are
provided to query this 'DNS' 'API'.
}
\author{
Bob Rudis (bob@rud.is)
}
