% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{print.cf_dns_result}
\alias{print.cf_dns_result}
\title{S3 Print method for cf_dns_result}
\usage{
\method{print}{cf_dns_result}(x, ...)
}
\arguments{
\item{x}{cf_dns_result}

\item{...}{ignored}
}
\description{
S3 Print method for cf_dns_result
}
