% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{valid_dns_types}
\alias{valid_dns_types}
\title{Valid DNS types}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 92 rows and 4 columns.}
\usage{
valid_dns_types
}
\description{
Valid DNS types
}
\keyword{datasets}
