% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\alias{dig}
\alias{nslookup}
\title{Make a single DNS query}
\usage{
query(name, type = "A")

dig(name, type = "A")

nslookup(name, type = "A")
}
\arguments{
\item{name}{query name (e.g. \code{example.com})}

\item{type}{query type (e.g. \code{AAAA}). Valid types can be found \href{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-4}{here}
and perused in the \link{valid_dns_types} built-in data frame. Defaults to "\code{A}". Can be alpha or numeric types.}
}
\value{
a `list`` with the following field meanings:
\itemize{
\item \code{Status}: The Response Code of the DNS Query. These are defined here: \url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-6}
\item \code{TC}: If true, it means the truncated bit was set. This happens when the DNS answer is larger than a single UDP or TCP packet. TC will almost always be false with Cloudflare DNS over HTTPS because Cloudflare supports the maximum response size.
\item \code{RD}: If true, it means the Recursive Desired bit was set. This is always set to true for Cloudflare DNS over HTTPS.
\item \code{RA}: If true, it means the Recursion Available bit was set. This is always set to true for Cloudflare DNS over HTTPS.
\item \code{AD}: If true, it means that every record in the answer was verified with DNSSEC.
\item \code{CD}: If true, the client asked to disable DNSSEC validation. In this case, Cloudflare will still fetch the DNSSEC-related records, but it will not attempt to validate the records.
\item \code{Question}: a data frame with the \code{name} and \code{type} (numeric form) from the original query
\item \code{Answer}: a data frame with the \code{name} of the record owner, \code{type} of record (numeric form), the \code{TTL} number of seconds the answer can be stored in cache before it is considered stale and the \code{answer} data (whhich is contingent on the question being asked)
}
}
\description{
The \code{\link[=query]{query()}} function is aliased to \code{\link[=dig]{dig()}} and \code{\link[=nslookup]{nslookup()}} since the
package author sometimes forgets they're at an R prompt and not a \code{bash} prompt.
}
\note{
Cloudfare \href{https://developers.cloudflare.com/1.1.1.1/nitty-gritty-details/}{does not support} \code{ANY} queries.
}
\examples{
query("rud.is")
query("example.com", "A")
query("example.com", "1")
query("microsoft.com", "MX")
# query("google-public-dns-a.google.com", "TXT") # CURRENTLY BROKEN ON CLOUDFLARE'S END
query("apple.com")
query("17.142.160.59", "PTR")
}
